requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.3
blis==0.7.11
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.3
cymem==2.0.8
dataclasses-json==0.6.1
dnspython==2.4.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
exceptiongroup==1.1.3
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.315
langcodes==3.3.0
langsmith==0.0.44
loguru==0.7.2
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
numpy==1.24.4
openai==0.28.1
packaging==23.2
pinecone-client==2.2.4
preshed==3.0.9
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.22
srsly==2.4.8
tenacity==8.2.3
thinc==8.2.1
tiktoken==0.5.1
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.6
wasabi==1.1.2
weasel==0.3.3
Werkzeug==3.0.0
yarl==1.9.2
zipp==3.17.0
